server/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-dotenv==1.0.0